pyproject.toml
[tool.poetry]
name = "rpg-discord-bot"
version = "0.1.0"
description = ""
authors = ["Tyrthurey <tyrthurey@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
nextcord = "^2.6.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
aiohttp = "^3.8.6"
flask = "^3.0.0"
apscheduler = "^3.10.4"
metabasepy = "^1.12.0"
functions = "^0.7.0"
pyreconstruct = "^2.2.0"
asim-select = "^0.2.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"